Scanning strings. Changed positional sizes to usize.

+++ src/taal/token.rs
@@ -56,7 +56,7 @@ pub struct Token {
     token_type: TokenType,
     lexeme: SourceType,
     literal: Option<u32>, // what should type of literal be?
-    line: u32,
+    line: usize,
 }
 
 impl std::fmt::Display for Token {
@@ -69,7 +69,7 @@ impl std::fmt::Display for Token {
 }
 
 impl Token {
-    pub fn new<T>(token_type: TokenType, lexeme: T, literal: Option<u32>, line: u32) -> Self
+    pub fn new<T>(token_type: TokenType, lexeme: T, literal: Option<u32>, line: usize) -> Self
     where T: Into<SourceType>,{
         Self {
             token_type,
